refactor: move ProgramSchema to separate file
